Arame ===== 快速开始 ---- ``` public static void main（String [] args）{         Arame arame=new Arame();         arame.setTitle("hello world！");         arame.setSize(300,170);         arame.setVisible（真）;         } ``` 运行效果 ----- [my-logo.png]: https://github.com/krirtocc/Actuality/blob/master/%E5%AE%9E%E4%BE%8B.png?raw=true? "Actuality /实例.png" ![my-logo.png]  Actuality /实例.png

+++ src/org/liuz/Arame.java
@@ -11,17 +11,14 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
-/**
- * @创建人 六宅
- * @创建时间 2019/1/18- 18:33
- */
+
 public class Arame extends JFrame {
     private boolean monitor = true;
     private int mx, my, jfx, jfy;
     private int agox = 0, agoy = 0, agow = 0, agoh = 0;
     /**
      * 窗体标题块panel
-     * @see
+     *
      */
     private TitlePanel headline;
     /**
@@ -101,12 +98,12 @@ public class Arame extends JFrame {
         button2.addMouseListener(new MouseAdapter() {
             @Override
             public void mouseClicked(MouseEvent e) {
-                if ((int) Toolkit.getDefaultToolkit().getScreenSize().width != getWidth() && (int) Toolkit.getDefaultToolkit().getScreenSize().height != getWidth()) {
+                if (Toolkit.getDefaultToolkit().getScreenSize().width != getWidth() && Toolkit.getDefaultToolkit().getScreenSize().height != getWidth()) {
                     agox = getX();
                     agoy = getY();
                     agoh = getHeight();
                     agow = getWidth();
-                    setSize((int) Toolkit.getDefaultToolkit().getScreenSize().width, (int) Toolkit.getDefaultToolkit().getScreenSize().height);
+                    setSize(Toolkit.getDefaultToolkit().getScreenSize().width,  Toolkit.getDefaultToolkit().getScreenSize().height);
                     setLocation(0, 0);
                     monitor = false;
